requirements.txt
fastapi
uvicorn
pyswisseph
jyotishganit
google-generativeai>=0.7.2
python-dotenv
sqlalchemy
skyfield         
psycopg2-binary
requests
supabase
pydantic
reportlab
python-multipart
yake
tenacity
openai